# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "description": "Returns a welcome message including the application’s configured name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Base"
                ],
                "summary": "Welcome message",
                "responses": {
                    "200": {
                        "description": "Welcome to App!"
                    }
                }
            }
        },
        "/api/v1/cron/control": {
            "post": {
                "description": "Start or stop the cron based on the \"action\" field.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cron"
                ],
                "summary": "Control cron job",
                "parameters": [
                    {
                        "description": "start or stop",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CronRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cron job started"
                    },
                    "202": {
                        "description": "Cron job will be stopped"
                    },
                    "400": {
                        "description": "Invalid request payload"
                    }
                }
            }
        },
        "/api/v1/list/sent-messages": {
            "get": {
                "description": "Retrieves all messages that have been sent.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "List sent messages",
                "responses": {
                    "200": {
                        "description": "Messages fetched successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Message"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to fetch messages"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Simple endpoint to verify the service is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Base"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Healthy!"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CronRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isSent": {
                    "type": "boolean"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        }
    }
}